pets/kirbochomp/requirements.txt
Adafruit-Blinka==8.52.0
adafruit-blinka-displayio==0.11.1
adafruit-circuitpython-bitmap_font==2.1.4
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-display-text==3.2.2
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-ticks==1.1.2
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinka-displayio-pygamedisplay==2.4.0
numpy==2.0.2
pillow==11.1.0
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
pyusb==1.3.1
typing_extensions==4.12.2
